Activado y mejorado el link de archivar

+++ app/Views/jugando.php
@@ -22,33 +22,47 @@
             </thead>
             <tbody>
                 <?php 
-                    foreach ($jugando as $key => $juego) {
+                    if ($jugando) {
+                        foreach ($jugando as $key => $juego) {
+                            echo '<tr>';
+                            echo '<td>'.$juego->juego.'</td>';
+                            echo '<td>'.$juego->sistema.'</td>';
+                            echo '<td>
+                                    <a type="button" id="btn-editar" href="'.site_url().'registra-partida/'.$juego->id.'" class="link">
+                                        '.date("Y-m-d || H:i", strtotime($juego->updated_at)).'
+                                    </a>
+                                </td>';
+                            echo '<td>
+                                    <a type="button" id="btn-editar" href="'.site_url().'archivar/'.$juego->id.'" class="edit">
+                                        <img src="'.site_url().'public/img/archivar.png" width="20" >
+                                    </a>
+                                </td>';
+                            echo '<td>
+                                    <a type="button" id="btn-editar" href="'.site_url().'terminar/'.$juego->id.'" class="edit">
+                                        <img src="'.site_url().'public/img/btndone.png" width="20" >
+                                    </a>
+                                </td>';
+                            echo '<td> 
+                                    <a type="button" id="btn-editar" href="'.site_url().'eliminar/'.$juego->id.'" class="edit">
+                                        <img src="'.site_url().'public/img/btncancel.png" width="20" >
+                                    </a>
+                                </td>';
+                            echo '</tr>';
+                        }
+                    
+                    } else {
                         echo '<tr>';
-                        echo '<td>'.$juego->juego.'</td>';
-                        echo '<td>'.$juego->sistema.'</td>';
-                        echo '<td>
-                                <a type="button" id="btn-editar" href="'.site_url().'registra-partida/'.$juego->id.'" class="link">
-                                    '.date("Y-m-d || H:i", strtotime($juego->updated_at)).'
-                                </a>
-                            </td>';
-                        echo '<td>
-                                <a type="button" id="btn-editar" href="'.site_url().'archivar/'.$juego->id.'" class="edit">
-                                    <img src="'.site_url().'public/img/archivar.png" width="20" >
-                                </a>
-                            </td>';
-                        echo '<td>
-                                <a type="button" id="btn-editar" href="'.site_url().'terminar/'.$juego->id.'" class="edit">
-                                    <img src="'.site_url().'public/img/btndone.png" width="20" >
-                                </a>
-                            </td>';
-                        echo '<td> 
-                                <a type="button" id="btn-editar" href="'.site_url().'eliminar/'.$juego->id.'" class="edit">
-                                    <img src="'.site_url().'public/img/btncancel.png" width="20" >
-                                </a>
-                            </td>';
+                        echo '<td>No hay datos disponibles</td>';
+                        echo '<td></td>';
+                        echo '<td></td>';
+                        echo '<td></td>';
+                        echo '<td></td>';
+                        echo '<td></td>';
                         echo '</tr>';
                     }
+
                 ?>
+                    
             </tbody>
         </table>
         </div>
